* Map::getTile() constraint checks to prevent undefined behaviour near map borders (x:0->20, y:0->20, x:65536->65536, y:65536->65536)

+++ otserv/trunk/map.h
@@ -185,7 +185,7 @@ public:
 	* Get a single tile.
 	* \return A pointer to that tile.
 	*/
-	Tile* getTile(uint16_t x, uint16_t y, uint16_t z);
+	Tile* getTile(int32_t x, int32_t y, int32_t z);
 	Tile* getTile(const Position& pos);
 
 	QTreeLeafNode* getLeaf(uint16_t x, uint16_t y){ return root.getLeaf(x, y);}
